buddy_list: Change all references to be specific to users.
This is a prep commit for adding a two section layout to the
buddy_list. It ensures that we can add a "other_keys" array without
breaking any of the existing "user_keys" functionality.

Note the addition of a TwoSectionListCursor class in list_cursor.js
which is a hack to ensure that list_cursor functionality works on the
"user_keys".

Note that this is more than just a "keys" to "user_keys" rename, we
also change the name of "find_li" to "find_user_li", "insert_new_html"
to "insert_new_html_for_user", "find_position" to "find_user_position"
and "insert_or_move" to "insert_or_move_user".