favicon: Use realm icon as browser favicon.
Previously, it was difficult to distinguish between multiple Zulip
realms, especially when pinned in the browser, as they all used the
same, default Zulip favicon. This commit allows users to set the
favicon to be the same as realm icon. (We still show unread counts
over the favicon, and the red dot in case of pms, in the same manner
as before).

There are a couple of ways to produce the base64 data url, some of
which include making the request to the url using XMLHttpRequest() or
converting to base64 via canvas.toDataURL(). We chose to use fetch()
and reader.result respectively.

Implementing this naively would have caused our clients to make a
fetch() every time their unread count is updated, but we avoid this by
storing the icon_data_url and simply using the stored value instead of
generating it repeatedly. This means that we also conditionally need
to regenerate the icon_data_url if the realm_icon is changed.

The possibility of generating the icon_data_url when the realm_icon is
uploaded, and making it available through eg page_params instead of
generating it when it's needed through fetch() (ie precomputing) still
exists, but it's not clear if it would be worth the cost.

This commit also completes the /help doc describing this user setting.

Fixes: #11752.