buddy_list: Rename references to "keys", "items", "users" and "ids".
This is a prep commit for enabling a two section layout in the
buddy_list. The sections will be for "users" and "others".

In this commit we rename:
"get_items_for_users" to "get_items_for_people",
"keys" to "user_keys",
"render_count" to "users_render_count",
and modify buddy_list.populate to expect "key_groups" or "{user_keys:
user_ids}" rather than "{keys: user_ids}" as well as make a similar
change wherever it is called.