settings: Add user display setting to enable realm icon as favicon.
This commit adds a setting in the user display settings, under the
advanced section which we will use to enable or disable the "realm
icon as favicon" feature.

It also changes zulip.yaml and creates the relevant help doc.

As the documentation mentions, this feature has not been implemented
yet, this is just the display setting plumbing.