recent_topics_ui: Delay resize_app to next animation frame via rAF cb.
Continuing the efforts to reduce dom trashing from the previous
commit, here we remove the second forced reflow by pushing the call to
navbar_alerts.resize_app() into the rAF callback queue.

This has no effect on the first renarrow, a ~ 90 ms decrease in the
second renarrow and a ~ 7 ms increase in the third renarrow, but
what's more important is that it eliminates one reflow and helps us
work to strong gains in subsequent commits.