buddy_list: Implement two section layout.
This commit implements a new feature which splits the buddy_list into
two sections, which the code internally refers to as the "users"
section and the "others" section, but for which we conditionally
display the titles "Subscribers"/"Recipients" and "Others".

This commit has several comments about changes that still need to be
made.

This commit also deletes the unused "chunk count" member of
buddy_list.