status_emoji: Bind animate handlers on `li.user_sidebar_entry`s.
Previously, there was a bug where mousing over
`li.user_sidebar_entry`s would not trigger the animation, as a result
of the fact that the `mouseenter` handler was bound to
`#user_presences` ie to the `ul`.

This commit fixes the above bug by binding the handlers to the
`li.user_sidebar_entry`s and rebinding when we remove a row.