stream_filter: Match streams by checking at word boundaries.
Previously, the stream filter would match stream names by spliting at
space and checking for prefix matches in the split strings, this would
result in two bugs:
1. Results for "bar baz" would not include "foo bar baz".
2. Results for "baz" would not include "foo bar/baz".

With (2) being more interesting, in that the separator could be a wide
variety of characters.

This commit changes the behaviour by implementing a method to filter
by matching at word boundaries.