emoji: Add emoji_animation_config user display setting.
We've received requests for the ability to completely turn off
animated emoji, because they can be distracting. Another reasonable
option to offer here is to only play animated emoji on hover. As such,
this commit adds a user display setting, emoji_animation_config, that
will be used in future commits.

The default value is "on_hover", this would be a change from the
existing behaviour which has animated emoji play on loop continuously
(ie what we're now going to call "always" behaviour).

The reason for changing the default behaviour is that continuously
playing any animation would have consequences to performance (and thus
also to power consumption).