message_emoji: Show still image for animated emoji based on setting.
Previously, we'd show animated emoji in messages in a continuous loop,
this had the potential to be annoying in the mild case or distracting
in the worst case. In order to prevent the above, we had added a
setting, emoji_animation_config, which would determine how animated
emojis behaved.

As such, this commit implements "on_hover" and "never" behaviour, and
wires up emoji_animation_config to determine which behaviour to use.

This commit essentially does to in message emoji what previous commits
have done to reaction and status emoji.